+++ app/services/import/gias_group_csv_importer_service.rb
@@ -1,4 +1,4 @@
-class Import::GiasGroupCsvImporterService
+class Import::GiasGroupCsvImporterService < Import::GiasCsvImporterService
   require "csv"
 
   IMPORT_MAP = {
@@ -18,8 +18,6 @@ class Import::GiasGroupCsvImporterService
   REQUIRED_VALUES = [
     :unique_group_identifier
   ].freeze
-
-  ENCODING = "ISO-8859-1"
 
   def initialize(path)
     @path = path
@@ -61,34 +59,6 @@ class Import::GiasGroupCsvImporterService
     import_result
   end
 
-  def required_column_headers_present?
-    file = File.open(@path, encoding: ENCODING)
-    headers = CSV.parse_line(file)
-    return false if headers.nil?
-
-    IMPORT_MAP.values.to_set.subset?(headers.to_set)
-  end
-
-  def changed_attributes(csv_attributes, model_attributes)
-    model_attribute_strings = model_attributes.transform_values(&:to_s)
-    csv_attribute_strings = csv_attributes.transform_values(&:to_s)
-
-    result = {}
-    csv_attribute_strings.each_pair do |key, value|
-      unless model_attribute_strings[key] == value
-        result[key] = {previous_value: model_attribute_strings[key], new_value: value}
-      end
-    end
-    result
-  end
-
-  def csv_row_attributes(row)
-    attributes = {}
-    IMPORT_MAP.each_pair do |key, value|
-      attributes[key.to_s] = row.field(value)
-    end
-    attributes
-  end
 
   def import_row(row)
     @csv_rows += 1
@@ -126,13 +96,6 @@ class Import::GiasGroupCsvImporterService
     end
 
     true
-  end
-
-  private def required_values_empty?(row)
-    values = REQUIRED_VALUES.map do |value|
-      row.field(IMPORT_MAP[value]).blank?
-    end
-    values.none?
   end
 
   private def import_result
