+++ spec/services/import/gias_headteacher_csv_importer_service_spec.rb
@@ -297,7 +297,7 @@ RSpec.describe Import::GiasHeadteacherCsvImporterService do
     end
   end
 
-  describe "#contact_csv_row_attributes" do
+  describe "#csv_row_attributes" do
     it "returns the values that are required by the importer and nothing else" do
       row = CSV::Row.new(
         ["URN", "Other column"],
@@ -305,7 +305,7 @@ RSpec.describe Import::GiasHeadteacherCsvImporterService do
       )
 
       service = described_class.new("/path")
-      row_attributes = service.contact_csv_row_attributes(row)
+      row_attributes = service.csv_row_attributes(row)
 
       expect(row_attributes["establishment_urn"]).to eql "123456"
       expect(row_attributes.keys).not_to include("Other column")
@@ -319,7 +319,7 @@ RSpec.describe Import::GiasHeadteacherCsvImporterService do
       )
 
       service = described_class.new("/path")
-      row_attributes = service.contact_csv_row_attributes(row)
+      row_attributes = service.csv_row_attributes(row)
 
       expect(row_attributes.keys).to include("establishment_urn")
       expect(row_attributes.keys).to include("name")
